Crate networks

Constants

BLOCK_DELAY_SECS
Time, in seconds, between each block.

DEFAULT_BOOTSTRAP
Bootstrap peer ids

DEFAULT_GENESIS
Default genesis car file bytes.

NEWEST_NETWORK_VERSION
Current network version for the network

UPGRADE_ACTORS_V2_HEIGHT
V4 network upgrade

UPGRADE_ACTORS_V3_HEIGHT
V10 network upgrade

UPGRADE_ACTORS_V4_HEIGHT
V12 network upgrade

UPGRADE_BREEZE_HEIGHT
V1 network upgrade

UPGRADE_CALICO_HEIGHT
V7 network upgrade

UPGRADE_CLAUS_HEIGHT
Remove burn on window PoSt fork

UPGRADE_IGNITION_HEIGHT
V3 network upgrade

UPGRADE_KUMQUAT_HEIGHT
V6 network upgrade

UPGRADE_LIFTOFF_HEIGHT
Switching to mainnet network name

UPGRADE_NORWEGIAN_HEIGHT
V11 network upgrade

UPGRADE_ORANGE_HEIGHT
V9 network upgrade

UPGRADE_PERSIAN_HEIGHT
V8 network upgrade

UPGRADE_PLACEHOLDER_HEIGHT

UPGRADE_SMOKE_HEIGHT
V2 network upgrade

UPGRADE_TAPE_HEIGHT
V5 network upgrade

Functions

beacon_schedule_default
Constructs a drand beacon schedule based on the build config.

get_network_version_default
Gets network version from epoch using default Mainnet schedule.